% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGLS-functions.R
\name{download_CGLS_data}
\alias{download_CGLS_data}
\title{Download CGLS data}
\usage{
download_CGLS_data(path, username, password, timeframe, product, resolution, version)
}
\arguments{
\item{path}{Working directory, for example: D:/CGLS_Data}

\item{username}{Register at https://land.copernicus.eu/global/}

\item{password}{Register at https://land.copernicus.eu/global/}

\item{timeframe}{Timeframe of interest (as daily date vector), for example june 2019: seq(as.Date("2019-06-01"), as.Date("2019-06-15"), by="days")}

\item{product}{Product name: chose from fapar, fcover, lai, ndvi,  ss, swi, lst, ...}

\item{resolution}{1km, 300m or 100m}

\item{version}{Version number: v1, v2, v3,...}
}
\value{
CGLS data Data saved locally in chosen folder.
}
\description{
Downloads manifest files of the Copernicus Global Land Service. Registration at https://land.copernicus.eu/global/ is required.
}
\details{
Check https://land.copernicus.eu/global/products/ for a product overview and product details. Check https://land.copernicus.vgt.vito.be/manifest/ for an overview for data availability in the manifest.
}
