% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGLS-functions.R
\name{download_CGLS_data}
\alias{download_CGLS_data}
\title{Download CGLS data}
\usage{
download_CGLS_data(
  path,
  username,
  password,
  timeframe,
  product,
  resolution,
  version
)
}
\arguments{
\item{path}{DESCRIBE PATH HERE}

\item{username}{DESCRIBE HERE}

\item{password}{DESCRIBE HERE}

\item{timeframe}{DESCRIBE HERE}

\item{product}{DESCRIBE HERE}

\item{resolution}{DESCRIBE HERE}

\item{version}{DESCRIBE HERE}
}
\value{
CGLS data DESCRIBE HERE
}
\description{
Download CGLS data
}
