% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGLS-functions.R
\name{nc_open_CGLS_data}
\alias{nc_open_CGLS_data}
\title{Open netcdf CGLS data}
\usage{
nc_open_CGLS_data(path, date, product, resolution, version)
}
\arguments{
\item{path}{DESCRIBE PATH HERE}

\item{date}{DESCRIBE HERE}

\item{product}{DESCRIBE HERE}

\item{resolution}{DESCRIBE HERE}

\item{version}{DESCRIBE HERE}
}
\value{
CGLS data DESCRIBE HERE
}
\description{
Open netcdf CGLS data
}
