% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGLS-functions.R
\name{ncvar_get_CGSL_data}
\alias{ncvar_get_CGSL_data}
\title{Read netcdf CGLS data}
\usage{
ncvar_get_CGSL_data(path, date, product, resolution, version, variable)
}
\arguments{
\item{path}{DESCRIBE PATH HERE}

\item{date}{DESCRIBE HERE}

\item{product}{DESCRIBE HERE}

\item{resolution}{DESCRIBE HERE}

\item{version}{DESCRIBE HERE}

\item{variable}{DESCRIBE HERE}
}
\value{
CGLS data DESCRIBE HERE
}
\description{
Read netcdf CGLS data
}
