% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCGLS-functions.R
\name{stack_CGLS_data}
\alias{stack_CGLS_data}
\title{stack CGLS data}
\usage{
stack_CGLS_data(path, timeframe, product, resolution, version, variable)
}
\arguments{
\item{path}{DESCRIBE PATH HERE}

\item{timeframe}{DESCRIBE HERE}

\item{product}{DESCRIBE HERE}

\item{resolution}{DESCRIBE HERE}

\item{version}{DESCRIBE HERE}

\item{variable}{DESCRIBE HERE}
}
\value{
CGLS data DESCRIBE HERE
}
\description{
stack CGLS data
}
